backend/requirements.txt
fastapi>=0.116.0
uvicorn[standard]>=0.35.0
langchain>=0.3.27
langchain-openai>=0.3.9
python-dotenv>=1.1.1
sqlalchemy[asyncio]>=2.0.40
asyncpg>=0.31.0
alembic>=1.16.0
bcrypt>=4.3.0
pyjwt>=2.10.0
pydantic[email]>=2.13.0
